Different defaults files for EL and Debian
Previously, this role tried to install /etc/sysconfig/rkhunter on
Debian, but the /etc/sysconfig directory does not even exist. Therefore,
I moved the Red Hat config file to its own task, and added a separate
task for /etc/default/rkhunter on Debian.

The file on Debian will be parameterized in a later commit. The original
is here to keep record of the default values.