Made the update tasks conditional
Assuming people run this role as part of config management, you do not
want to run the `propupd` and `update` tasks every time. In fact, the
cronjob should do that for you.

Therefore, I made these tasks conditional on the existence of the
database files created by `propupd` and `update`.